1. bugfix list surah viewModel
